Extract erlavro decoder options to separate module
